+++ src/DependencyInjection/Configuration.php
@@ -16,13 +16,6 @@ class Configuration implements ConfigurationInterface
     public const DATABASE_URL = 'database_url';
     public const DEFAULT_BLOB_TYPE = 'default_blob_type';
     public const MAY_CREATE_FORM = 'MAY_CREATE_FORM';
-    public const MAY_UPDATE_FORM = 'MAY_UPDATE_FORM';
-    public const MAY_DELETE_FORM = 'MAY_DELETE_FORM';
-    public const MAY_READ_FORM = 'MAY_READ_FORM';
-    public const MAY_CREATE_FORM_SUBMISSIONS = 'MAY_CREATE_FORM_SUBMISSIONS';
-    public const MAY_UPDATE_FORM_SUBMISSIONS = 'MAY_UPDATE_FORM_SUBMISSIONS';
-    public const MAY_DELETE_FORM_SUBMISSIONS = 'MAY_DELETE_FORM_SUBMISSIONS';
-    public const MAY_READ_FORM_SUBMISSIONS = 'MAY_READ_FORM_SUBMISSIONS';
 
     /**
      * @throws BlobApiError
@@ -50,20 +43,6 @@ class Configuration implements ConfigurationInterface
         return AuthorizationConfigDefinition::create()
             ->addResourcePermission(self::MAY_CREATE_FORM, 'false',
                 'Returns true if the user is allowed to add the given form, false otherwise')
-            ->addResourcePermission(self::MAY_UPDATE_FORM, 'false',
-                'Returns true if the user is allowed to update the given form, false otherwise')
-            ->addResourcePermission(self::MAY_DELETE_FORM, 'false',
-                'Returns true if the user is allowed to delete the given form, false otherwise')
-            ->addResourcePermission(self::MAY_READ_FORM, 'false',
-                'Returns true if the user is allowed to read the given form, false otherwise')
-            ->addResourcePermission(self::MAY_CREATE_FORM_SUBMISSIONS, 'false',
-                'Returns true if the user is allowed to add a submission to the given form, false otherwise')
-            ->addResourcePermission(self::MAY_UPDATE_FORM_SUBMISSIONS, 'false',
-                'Returns true if the user is allowed to update all form submissions, false otherwise')
-            ->addResourcePermission(self::MAY_DELETE_FORM_SUBMISSIONS, 'false',
-                'Returns true if the user is allowed to delete all form submissions, false otherwise')
-            ->addResourcePermission(self::MAY_READ_FORM_SUBMISSIONS, 'false',
-                'Returns true if the user is allowed to read all form submissions, false otherwise')
             ->getNodeDefinition();
     }
 }